+++ include/zephyr/arch/riscv/arch.h
@@ -16,7 +16,7 @@
 #define ZEPHYR_INCLUDE_ARCH_RISCV_ARCH_H_
 
 #include <zephyr/arch/riscv/thread.h>
-#include <zephyr/arch/riscv/exp.h>
+#include <zephyr/arch/riscv/exception.h>
 #include <zephyr/arch/riscv/irq.h>
 #include <zephyr/arch/riscv/sys_io.h>
 #include <zephyr/arch/common/sys_bitops.h>
@@ -29,7 +29,7 @@
 #include <soc.h>
 #include <zephyr/devicetree.h>
 #include <zephyr/arch/riscv/csr.h>
-#include <zephyr/arch/riscv/exp.h>
+#include <zephyr/arch/riscv/exception.h>
 
 /* stacks, for RISCV architecture stack should be 16byte-aligned */
 #define ARCH_STACK_PTR_ALIGN  16
