+++ include/zephyr/arch/riscv/exception.h
